# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mosaic API",
    "description": "Backend API service for Mosaic temporary digital spaces",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/spaces": {
      "post": {
        "tags": ["Spaces"],
        "summary": "Create Space",
        "operationId": "create_space_api_v1_spaces_post",
        "parameters": [
          {
            "name": "mosaic_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mosaic Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Spaces"],
        "summary": "List My Spaces",
        "operationId": "list_my_spaces_api_v1_spaces_get",
        "parameters": [
          {
            "name": "mosaic_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mosaic Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpaceSummary"
                  },
                  "title": "Response List My Spaces Api V1 Spaces Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/spaces/{space_id}/activity": {
      "get": {
        "tags": ["Activity"],
        "summary": "Get Space Activity",
        "operationId": "get_space_activity_api_v1_spaces__space_id__activity_get",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Fetch events occurred strictly after this sequence",
              "default": 0,
              "title": "After Sequence"
            },
            "description": "Fetch events occurred strictly after this sequence"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max events to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max events to return"
          },
          {
            "name": "mosaic_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mosaic Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityFeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityEventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "space_id": {
            "type": "string",
            "title": "Space Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "actor": {
            "$ref": "#/components/schemas/EventActorInfo"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "space_id",
          "sequence",
          "type",
          "occurred_at",
          "actor"
        ],
        "title": "ActivityEventResponse"
      },
      "ActivityFeedResponse": {
        "properties": {
          "space_id": {
            "type": "string",
            "title": "Space Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ActivityEventResponse"
            },
            "type": "array",
            "title": "Events"
          },
          "latest_sequence": {
            "type": "integer",
            "title": "Latest Sequence"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": ["space_id", "events", "latest_sequence", "has_more"],
        "title": "ActivityFeedResponse"
      },
...
      "CreateSpaceRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Title"
          },
          "template": {
            "type": "string",
            "title": "Template",
            "description": "Template/purpose: trip, event, team, gathering, custom",
            "default": "gathering"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "cover_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Color",
            "default": "#246A5A"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          }
        },
        "type": "object",
        "required": ["title"],
        "title": "CreateSpaceRequest"
      },
...
      "EventActorInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "avatar_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Asset Id"
          }
        },
        "type": "object",
        "required": ["id", "display_name"],
        "title": "EventActorInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpaceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "cover_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Asset Id"
          },
          "cover_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Color"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "current_role": {
            "type": "string",
            "title": "Current Role"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "template",
          "status",
          "created_by",
          "created_at",
          "current_role"
        ],
        "title": "SpaceResponse"
      },
      "SpaceSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "cover_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Color"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "current_role": {
            "type": "string",
            "title": "Current Role"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "template",
          "status",
          "created_at",
          "current_role",
          "member_count"
        ],
        "title": "SpaceSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}